part5

+++ app/src/main/java/com/example/todolistapp/Database/TaskDAO.kt
@@ -10,10 +10,10 @@ interface TaskDAO {
     suspend fun insertTask(task: Task): Long
 
     @Update
-    suspend fun updateTask(task: Task): Long
+    suspend fun updateTask(task: Task): Int
 
     @Delete
-    suspend fun deleteTask(task: Task): Long
+    suspend fun deleteTask(task: Task): Int
 
     @Query( "DELETE FROM task_data_table")
     suspend fun deleteAll()
